Removed the authenticated middleware. We'll need to figure out a new way to render harvester pages.
