+++ shell/components/templates/home.vue
@@ -9,6 +9,7 @@ import AzureWarning from '@shell/components/auth/AzureWarning';
 import BrowserTabVisibility from '@shell/mixins/browser-tab-visibility';
 import Inactivity from '@shell/components/Inactivity';
 import { mapState, mapGetters } from 'vuex';
+import Authenticated from '@shell/mixins/authenticated';
 
 export default {
 
@@ -21,9 +22,7 @@ export default {
     Inactivity
   },
 
-  mixins: [Brand, BrowserTabVisibility],
-
-  middleware: ['authenticated'],
+  mixins: [Brand, Authenticated, BrowserTabVisibility],
 
   data() {
     return {
