+++ shell/config/middleware.js
@@ -1,6 +1,3 @@
 const middleware = {};
 
-middleware['authenticated'] = require('../middleware/authenticated.js');
-middleware['authenticated'] = middleware['authenticated'].default || middleware['authenticated'];
-
 export default middleware;
